fix: harden rendering against null/destroyed objects, frame overflow, and concurrent dictionary mutation
- Render loop: skip null/destroyed objects before Draw() call
- GameSceneDrawingSorting: cache TNext at loop start in all 4 linked-list traversals so concurrent removal can't stale the iterator
- Mobile/Item: remove silent (byte) cast on animation frameIndex clamp — avoids index wrap when frame count > 255
- World, HealthLinesManager, AutoLootManager: snapshot Mobiles/Items.Values with .ToList() before iteration to prevent InvalidOperationException from network-thread modifications

+++ src/ClassicUO.Client/Game/GameObjects/Item.cs
@@ -517,7 +517,7 @@ namespace ClassicUO.Game.GameObjects
 
             if (LastAnimationChangeTime < Time.Ticks)
             {
-                byte frameIndex = (byte)(AnimIndex + (ExecuteAnimation ? 1 : 0));
+                int frameIndex = AnimIndex + (ExecuteAnimation ? 1 : 0);
                 ushort id = GetGraphicForAnimation();
 
                 bool mirror = false;
@@ -550,7 +550,7 @@ namespace ClassicUO.Game.GameObjects
                         // when the animation is done, stop to animate the corpse
                         if (frameIndex >= frames.Length)
                         {
-                            frameIndex = (byte)(frames.Length - 1);
+                            frameIndex = frames.Length - 1;
                         }
 
                         AnimIndex = (byte)(frameIndex % frames.Length);
